requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
requests==2.*

# Gemini
google-generativeai>=0.7.2
